# Speed up `is_homogeneous`