Some measurements taking the example from the `is_homogeneous` docstring:

    R, (x, y, z) = graded_polynomial_ring(QQ, ["x", "y", "z"], [1, 2, 3])
    f = x^2+y*z
    W = [1 2 1 0; 3 4 0 1]
    S, (w, x, y, z) = graded_polynomial_ring(QQ, ["w", "x", "y", "z"], W)
    F = w^3*y^3*z^3 + w^2*x*y^2*z^2 + w*x^2*y*z + x^3

Before:

    julia> @btime is_homogeneous(f);
      1.533 μs (62 allocations: 2.28 KiB)

    julia> @btime is_homogeneous(F);
      3.932 μs (152 allocations: 5.53 KiB)

After:

    julia> @btime is_homogeneous(f);
      314.970 ns (6 allocations: 288 bytes)

    julia> @btime is_homogeneous(F);
      642.638 ns (8 allocations: 520 bytes)


I will make a separate PR for Hecke which provides `addmul!`, `zero!` and more for `FinGenAbGroupElem`.

Also I think there are more places that could be similarly improved, e.g. `homogeneous_components`. However, unlike `is_homogeneous` they did not stick out in some timing profiles on CI tests we made this week.