@lgoettgens 2024-09-21
Maybe we wait for a release of https://github.com/thofma/Hecke.jl/pull/1619 and then can simplify the comments away?