@fingolfin 2024-09-21
But that then also needs to go into a new Hecke release, and then more benchmarking should be done to verify it doesn't get slower again etc. I'd rather have this in fact sooner. IMHO the "simplification" is not as important as the speed up.