@lgoettgens 2024-09-23
Hecke 0.34.4 is now released with the required changes.