@lgoettgens 2024-09-23
I now adapted to the new Hecke functions, and verified that this does not increase the allocations nor the runtime of the example above.

@fingolfin if you are happy with this, please merge